+++ changelog.md
@@ -10,6 +10,7 @@
 * Added `duration` property to `KymoTrack` which returns the duration (in seconds) that the track was observed.
 * Added option to filter tracks by duration in seconds using `lk.filter_tracks(tracks, minimum_duration=duration_in_seconds)`.
 * Added `KymoTrackGroup.filter()` to filter tracks in-place. `tracks.filter(minimum_duration=2)` is equivalent to `tracks = lk.filter_tracks(tracks, minimum_duration=2)`.
+* Added option to align plots vertically by passing `vertical=True` to `Scan.plot_correlated` and `ImageStack.plot_correlated()`.
 
 #### Bug fixes